requirements.txt
opencv-python-headless
numpy
onnxruntime
Pillow
python-bidi